Updates to better normalize JSON output
We have been erroneously mapping cpu-load to cpu-load-all when they are
actually two parallel output forms. This patch now handles both,
removing attempts to map cpu-load to cpu-load-all.

The rxbyt and txbyt fields were not being normalized to the newer names,
rxkB and txkB, with the values converted to kB. This patch performs that
mapping.

We also attempt to improve the performance of xmltojson, though it is
not conclusive. It seems a bit faster to not emit so many tabs between
fields, we have changed the testing to ensure we don't reintroduce them.